Blocks and branches

See merge request edmund/starkify/starkify!32

+++ testfiles/branch.c
@@ -1,12 +1,12 @@
-__attribute__((noinline)) unsigned int then_branch() {
+__attribute__((noinline)) unsigned short then_branch() {
   return 3;
 }
 
-__attribute__((noinline)) unsigned int else_branch() {
+__attribute__((noinline)) unsigned short else_branch() {
   return 4;
 }
 
-__attribute__((noinline)) unsigned int if_test(unsigned int i)
+__attribute__((noinline)) unsigned short if_test(unsigned short i)
 {
   if (i == 2)
     return then_branch();
@@ -14,4 +14,4 @@ __attribute__((noinline)) unsigned int if_test(unsigned int i)
     return else_branch();
 }
 
-unsigned int main() { return if_test(2) + if_test(5); }
+unsigned short main() { return if_test(2) + if_test(5); }
