+++ testfiles/rem.c
@@ -1,6 +1,6 @@
 __attribute__((noinline))
-unsigned int rem_u(unsigned int u, unsigned int m) {
+unsigned short rem_u(unsigned short u, unsigned short m) {
   return u % m;
 }
 
-unsigned int main() { return rem_u(100, 13); }
+unsigned short main() { return rem_u(100, 13); }
